Move AddrFamily to own module
